+++ examples/snowball.js
@@ -11,10 +11,7 @@ const bot = mineflayer.createBot({
 bot.loadPlugin(minecraftHawkEye)
 
 bot.on('spawn', function () {
-  bot.chat('/kill @e[type=minecraft:arrow]')
-  // bot.chat(`/give ${bot.username} crossbow{Enchantments:[{id:unbreaking,lvl:3}]} 1`)
-  bot.chat(`/give ${bot.username} crossbow{Enchantments:[{id:quick_charge,lvl:3},{id:unbreaking,lvl:100}]} 1`)
-  bot.chat(`/give ${bot.username} minecraft:arrow 300`)
+  bot.chat(`/give ${bot.username} minecraft:snowball 300`)
   bot.chat('/time set day')
   bot.chat('Ready!')
   fire()
@@ -24,27 +21,23 @@ bot.on('die', () => {
   bot.hawkEye.stop()
 })
 
-async function fire() {
-  // bot.chat('/kill @e[type=minecraft:arrow]')
-
-  const slotID = bot.getEquipmentDestSlot('hand')
-
+async function fire () {
+  /*
   if (bot.inventory.slots[slotID] === null || bot.inventory.slots[slotID].name !== 'crossbow') {
     const crossbow = bot.inventory.items().find(item => item.name.includes('crossbow'))
     if (crossbow) {
       await bot.equip(crossbow, 'hand')
     }
   }
+  */
 
   const target = bot.hawkEye.getPlayer() // Fire to nearest player
-  if (target && bot.inventory.slots[slotID] !== null && bot.inventory.slots[slotID].name === 'crossbow') {
-    // console.log(bot.entity.position.distanceTo(target.position))
-    // bot.hawkEye.oneShot(target, 'crossbow')
-    bot.hawkEye.autoAttack(target, 'crossbow')
+  if (target) {
+    bot.hawkEye.autoAttack(target, 'snowball')
     console.log('shooting')
   } else {
     setTimeout(() => {
       fire()
-    }, 500);
+    }, 500)
   }
 }